Moved AST rewriting to be done together with Normalization
