+++ go/vt/vtgate/engine/primitive.go
@@ -37,10 +37,6 @@ const (
 	// This is used for sending different IN clause values
 	// to different shards.
 	ListVarName = "__vals"
-	//LastInsertIDName is a reserved bind var name for last_insert_id()
-	LastInsertIDName = "__lastInsertId"
-	//DBVarName is a reserved bind var name for database()
-	DBVarName = "__vtdbname"
 )
 
 // VCursor defines the interface the engine will use
