+++ go/vt/vtgate/planbuilder/select.go
@@ -20,8 +20,6 @@ import (
 	"errors"
 	"fmt"
 
-	"vitess.io/vitess/go/vt/vterrors"
-
 	"vitess.io/vitess/go/vt/sqlparser"
 	"vitess.io/vitess/go/vt/vtgate/engine"
 )
@@ -126,11 +124,7 @@ func (pb *primitiveBuilder) processSelect(sel *sqlparser.Select, outer *symtab) 
 // pushes it down, and updates the route info if the new constraint improves
 // the primitive. This function can push to a WHERE or HAVING clause.
 func (pb *primitiveBuilder) pushFilter(in sqlparser.Expr, whereType string) error {
-	rewritten, err := RewriteAndUpdateBuilder(in, pb)
-	if err != nil {
-		return vterrors.Wrap(err, "failed to Rewrite expressions")
-	}
-	filters := splitAndExpression(nil, rewritten)
+	filters := splitAndExpression(nil, in)
 	reorderBySubquery(filters)
 	for _, filter := range filters {
 		pullouts, origin, expr, err := pb.findOrigin(filter)
@@ -195,16 +189,7 @@ func (pb *primitiveBuilder) pushSelectRoutes(selectExprs sqlparser.SelectExprs) 
 	for _, node := range selectExprs {
 		switch node := node.(type) {
 		case *sqlparser.AliasedExpr:
-			rewritten, err := RewriteAndUpdateBuilder(node.Expr, pb)
-			if err != nil {
-				return nil, vterrors.Wrap(err, "failed to Rewrite expression")
-			}
-			if rewritten != node.Expr && node.As.IsEmpty() {
-				buf := sqlparser.NewTrackedBuffer(nil)
-				node.Expr.Format(buf)
-				node.As = sqlparser.NewColIdent(buf.String())
-			}
-			pullouts, origin, expr, err := pb.findOrigin(rewritten)
+			pullouts, origin, expr, err := pb.findOrigin(node.Expr)
 			if err != nil {
 				return nil, err
 			}
